Combine td and th styles
This is a minor change that combines a few styles shared between td and
th elements. This adds base styles to both elements and then overrides
only the 3 items that make the th unique (uppercase, bottom-aligned,
and reduced top/bottom padding).

From my point of view, this makes it easier to see which styles are
shared and should also simplify future updates to the styles of these
elements.