Set max content width
This change adjusts the styles to enforce a maximum width on the
content (currently 960px). To me, this looks better on wide displays
and is generally easier to work with. The hymn pages still look a
little off, as their content is often quite narrow, but this is
still an improvement to my eye over forcing content all the way to
the left.