Combine scripture and lyrics scraping (#100)
This change merges lyrics and scripture acquisition. Because we are already loading hymn pages, it saves requests and time to pull both of these fields at the same time rather than pulling lyrics from a separate API. This does update the lyrics used to match default the text used by hymnary.org and we now include the source hymnal with the text.